fix(ci): mypy numpy récent (convolve→astype) + build llama.cpp déterministe (web-UI OFF)
Deux échecs CI sur la release v0.3.4 :
- mypy (tests, main) : les stubs numpy d'une install fraîche typent
  np.convolve en floating[Any] — cast .astype(np.float64) explicite
  (acoustic_metrics), invisible avec le numpy épinglé du venv local
- publish-image (tag) : llama.cpp b9816 télécharge les assets de sa web-UI
  au build (ui-assets.cmake) et le bucket HF a répondu en erreur → CMake
  Error. On ne se sert de llama-server qu'en API : -DLLAMA_BUILD_UI=OFF
  -DLLAMA_USE_PREBUILT_UI=OFF dans les deux Dockerfiles (build déterministe,
  plus aucun fetch réseau pour l'UI)

Le tag v0.3.4 est déplacé sur ce commit (l'image du tag n'avait jamais été
publiée — pipeline rouge à la création).

+++ transcria/audio/acoustic_metrics.py
@@ -90,7 +90,8 @@ def estimate_rt60(
         return None
     power_db = 10.0 * np.log10(power + 1e-12)
     if power_db.size >= 3:
-        power_db = np.convolve(power_db, np.ones(3) / 3.0, mode="same")
+        # astype : les stubs numpy récents typent np.convolve en floating[Any] (vu en CI).
+        power_db = np.convolve(power_db, np.ones(3) / 3.0, mode="same").astype(np.float64)
     hop_s = hop / sample_rate
 
     # Parcours : chaque décroissance va d'un maximum local au minimum suivant.
